+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Makiko changelog
 
-## Unreleased
+## 0.2.4 (2024-11-09)
 
 - Add support for `diffie-hellman-group1-sha1` key exchange under `insecure-crypto` feature.
 - Add support for `3des-cbc` cipher under `insecure-crypto` feature.